+++ package.json
@@ -15,6 +15,7 @@
     "storybook:composed": "sleep 5 && yarn workspace storybooks storybook --quiet",
     "storybook:cr": "yarn workspace @guardian/common-rendering storybook --no-open --quiet",
     "storybook:dcr": "yarn workspace @guardian/dotcom-rendering storybook --no-open --quiet",
+    "storybook:ar": "cd apps-rendering && yarn storybook --no-open --quiet",
     "build-storybook": "yarn workspace storybooks build-storybook && yarn workspace @guardian/common-rendering build-storybook && yarn workspace @guardian/dotcom-rendering build-storybook",
     "build:dcr": "cd ./dotcom-rendering && yarn makeBuild",
     "chromatic": "chromatic --build-script-name=build-storybook --exit-zero-on-changes",
