+++ storybooks/Readme.stories.jsx
@@ -1,0 +1,12 @@
+const ReadMe = () => (
+	<div>
+		These are all the storybooks for the individual projects that live in
+		this repo.
+	</div>
+);
+
+export default {
+	component: ReadMe,
+};
+
+export const Readme = () => <ReadMe />;
