+++ storybooks/readme.mdx
@@ -1,0 +1,3 @@
+# DCR / AR / CR storybooks
+
+This project collates each of the storybooks in this repo.
